+++ clkmgr/proxy/connect_chrony.hpp
@@ -11,13 +11,15 @@
 
 #include "common/util.hpp"
 
+#include <cstddef>
+
 __CLKMGR_NAMESPACE_BEGIN
 
 class ConnectChrony
 {
   public:
     static void connect_chrony();
-    static int subscribe_chrony(int timeBaseIndex, sessionId_t sessionId);
+    static int subscribe_chrony(size_t timeBaseIndex, sessionId_t sessionId);
     static int remove_chrony_subscriber(sessionId_t sessionId);
 };
